Fix dark notification filter labels
Use the active theme's secondary text token in the animated category labels instead of the hard-coded light-theme color.
